requirements.txt
pyqt5==5.10.1
astropy==3.0.1
Send2Trash==1.5.0
Pillow==5.1.0
